+++ resources/lang/en/novaposhta.php
@@ -45,6 +45,7 @@ return [
         //3275 на 2020-11-09
         '20002503517' => 'The document does not belong to this user',
         '20000303942' => 'EN is located in the department of the "Receipt-issuance point" category, where there is no access to the IS. Ordering additional services is prohibited.',
+        '20003603881' => 'No orders found',
 
         '20000100004' => 'User with email exists',
         '20000100002' => 'General error',